+++ webknossos/tests/dataset/test_dataset_add_remote_mag_and_layer.py
@@ -40,7 +40,7 @@ def test_add_remote_mags_from_mag_view(
     sample_remote_dataset: Dataset,
     sample_layer_and_mag_name: Iterable[tuple[str, str]],
 ) -> None:
-    remote_dataset = RemoteDataset.open("l4_sample", "Organization_X")
+    remote_dataset = RemoteDataset.open("l4_sample", organization_id="Organization_X")
     sample_remote_mags = [
         remote_dataset.get_layer(layer).get_mag(mag)
         for layer, mag in sample_layer_and_mag_name
@@ -68,7 +68,7 @@ def test_add_remote_mags_from_path(
     sample_remote_dataset: Dataset,
     sample_layer_and_mag_name: Iterable[tuple[str, str]],
 ) -> None:
-    remote_dataset = RemoteDataset.open("l4_sample", "Organization_X")
+    remote_dataset = RemoteDataset.open("l4_sample", organization_id="Organization_X")
     sample_remote_mags = [
         remote_dataset.get_layer(layer).get_mag(mag)
         for layer, mag in sample_layer_and_mag_name
@@ -95,7 +95,7 @@ def test_add_remote_mags_from_path(
 
 
 def test_ref_layer_from_remote_layer(sample_remote_dataset: Dataset) -> None:
-    remote_dataset = RemoteDataset.open("l4_sample", "Organization_X")
+    remote_dataset = RemoteDataset.open("l4_sample", organization_id="Organization_X")
     assert remote_dataset.zarr_streaming_path is not None, (
         "Zarr streaming sets a remote path."
     )
@@ -116,7 +116,7 @@ def test_ref_layer_non_public(tmp_upath: UPath) -> None:
     dataset = Dataset.open("testdata/simple_zarr3_dataset").copy_dataset(
         tmp_upath / "simple_zarr3_dataset"
     )
-    remote_dataset = RemoteDataset.open("l4_sample", "Organization_X")
+    remote_dataset = RemoteDataset.open("l4_sample", organization_id="Organization_X")
     remote_dataset.is_public = False
     dataset.add_layer_as_ref(remote_dataset.get_layer("segmentation"), "segmentation")
 
@@ -131,7 +131,7 @@ def test_ref_layer_non_public(tmp_upath: UPath) -> None:
 
 def test_shallow_copy_remote_layers(tmp_upath: UPath) -> None:
     dataset = Dataset(tmp_upath / "origin", voxel_size=(10, 10, 10))
-    remote_dataset = RemoteDataset.open("l4_sample", "Organization_X")
+    remote_dataset = RemoteDataset.open("l4_sample", organization_id="Organization_X")
     dataset.add_layer_as_ref(remote_dataset.get_layer("color"), "color")
     dataset_copy = dataset.shallow_copy_dataset(tmp_upath / "copy")
     data = dataset_copy.get_layer("color").get_mag("16-16-4").read()