+++ webknossos/webknossos/cli/upload.py
@@ -69,12 +69,12 @@ def main(
     """Upload a dataset to a WEBKNOSSOS server."""
 
     with webknossos_context(url=webknossos_url, token=token):
-        folder_id: None | RemoteFolder = None
+        folder_obj: None | RemoteFolder = None
         if folder is not None:
-            folder_id = RemoteFolder.get_by_path(folder)
+            folder_obj = RemoteFolder.get_by_path(folder)
         Dataset.open(dataset_path=source).upload(
             new_dataset_name=dataset_name,
             jobs=jobs,
-            folder_id=folder_id,
+            folder=folder_obj,
             transfer_mode=transfer_mode,
         )