+++ webknossos/webknossos/dataset/dataset.py
@@ -614,12 +614,13 @@ class Dataset(AbstractDataset[Layer, SegmentationLayer]):
         *,
         new_dataset_name: str | None = None,
         initial_team_ids: list[str] | None = None,
-        folder_id: str | RemoteFolder | None = None,
+        folder: str | RemoteFolder | None = None,
         require_unique_name: bool = False,
         layers_to_link: Sequence[LayerToLink | RemoteLayer] | None = None,
         transfer_mode: TransferMode = TransferMode.HTTP,
         jobs: int | None = None,
         common_storage_path_prefix: str | None = None,
+        folder_id: str | RemoteFolder | None = None,
     ) -> "RemoteDataset":
         """Upload this dataset to webknossos.
 
@@ -629,12 +630,13 @@ class Dataset(AbstractDataset[Layer, SegmentationLayer]):
         Args:
             new_dataset_name: Name for the new dataset defaults to the current name.
             initial_team_ids: Optional list of team IDs to grant initial access
-            folder_id: Optional ID of folder where dataset should be placed
+            folder: Optional ID of folder where dataset should be placed
             require_unique_name: Whether to make request fail in case a dataset with the name already exists
             layers_to_link: Optional list of LayerToLink to link already published layers to the dataset.
             jobs: Optional number of jobs to use for uploading the data.
             common_storage_path_prefix: Optional path prefix used when transfer_mode is either COPY or MOVE_AND_SYMLINK
                                         to select one of the available WEBKNOSSOS storages.
+            folder_id: Deprecated, use folder instead.
         Returns:
             RemoteDataset: Reference to the newly created remote dataset
         Note:
@@ -660,8 +662,16 @@ class Dataset(AbstractDataset[Layer, SegmentationLayer]):
 
         new_dataset_name = new_dataset_name or self.name
 
-        if isinstance(folder_id, RemoteFolder):
-            folder_id = folder_id.id
+        if folder_id is not None:
+            warn_deprecated("folder_id", "folder")
+            if folder is not None:
+                raise ValueError(
+                    "Both folder_id and folder were provided. Only one is allowed."
+                )
+            folder = folder_id
+
+        if isinstance(folder, RemoteFolder):
+            folder = folder.id
 
         converted_layers_to_link = (
             None
@@ -682,7 +692,7 @@ class Dataset(AbstractDataset[Layer, SegmentationLayer]):
                 ApiReserveDatasetUplaodToPathsParameters(
                     dataset_name=new_dataset_name,
                     initial_team_ids=initial_team_ids or [],
-                    folder_id=folder_id,
+                    folder_id=folder,
                     require_unique_name=require_unique_name,
                     data_source=self._properties,
                     layers_to_link=[
@@ -720,7 +730,7 @@ class Dataset(AbstractDataset[Layer, SegmentationLayer]):
                 new_dataset_name,
                 converted_layers_to_link,
                 jobs,
-                folder_id=folder_id,
+                folder_id=folder,
             )
 
         return RemoteDataset.open(dataset_id=new_dataset_id)